requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
requests==2.32.3
pydantic==1.10.17
sqlparse==0.5.1
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
openai==1.43.0

